Split up types.mjs
